Backend/requirements.txt
streamlit
python-dotenv
requests
python-pptx
docx2txt
PyPDF2
langchain
langchain-text-splitters
numpy
